Renamed push to send, pull to bring and merged fetch into bring.
Cleaned up warnings in bring
