+++ storage/src/storage/mod.rs
@@ -391,7 +391,7 @@ impl From<Uuid> for XvcStorageGuid {
 
 pub const XVC_STORAGE_GUID_FILENAME: &str = ".xvc-guid";
 
-pub fn get_remote_from_store(
+pub fn get_storage_record(
     output_snd: Sender<XvcOutputLine>,
     xvc_root: &XvcRoot,
     identifier: &StorageIdentifier,
